Add a BeamMixinClass; add beam and beams parameters into the mixin classes
