+++ spectral_cube/spectral_cube.py
@@ -42,7 +42,7 @@ from .lower_dimensional_structures import (Projection, Slice, OneDSpectrum,
 from .base_class import (BaseNDClass, SpectralAxisMixinClass,
                          DOPPLER_CONVENTIONS, SpatialCoordMixinClass,
                          MaskableArrayMixinClass, MultiBeamMixinClass,
-                         HeaderMixinClass
+                         HeaderMixinClass, BeamMixinClass
                         )
 from .utils import (cached, warn_slow, VarianceWarning, BeamAverageWarning,
                     UnsupportedIterationStrategyWarning, WCSMismatchWarning,
@@ -3213,7 +3213,7 @@ class BaseSpectralCube(BaseNDClass, MaskableArrayMixinClass,
         return self._new_cube_with(data=dsarr, wcs=newwcs,
                                    mask=BooleanArrayMask(mask, wcs=newwcs))
 
-class SpectralCube(BaseSpectralCube):
+class SpectralCube(BaseSpectralCube, BeamMixinClass):
 
     __name__ = "SpectralCube"
 
@@ -3376,12 +3376,8 @@ class VaryingResolutionSpectralCube(BaseSpectralCube, MultiBeamMixinClass):
             raise ValueError("Beam list must have same size as spectral "
                              "dimension")
 
-        self._beams = beams
+        self.beams = beams
         self.beam_threshold = beam_threshold
-
-    @property
-    def beams(self):
-        return self._beams
 
     def __getitem__(self, view):
